Remove Binding initializers
