--- mvvm/src/main/java/com/thefuntasty/mvvm/binding/DataBindingVariables.kt
@@ -1,3 +1,0 @@
-package com.thefuntasty.mvvm.binding
-
-data class DataBindingVariables(val viewId: Int, val viewModelId: Int, val viewStateId: Int)
